Cast SUM result to match rewards_close_month return signature
Postgres widens SUM of bigint to numeric and SUM of numeric stays numeric,
so the RETURN QUERY columns now explicitly cast back to bigint and numeric
to match the OUT parameter types.

+++ supabase/migrations/011_rewards_monthly_payout.sql
@@ -71,9 +71,9 @@ BEGIN
 
   RETURN QUERY
   SELECT
-    COUNT(*)::int                                AS out_referrers_paid,
-    COALESCE(SUM((c.total_zar * 100)::bigint), 0) AS out_points_awarded,
-    COALESCE(SUM(c.total_zar), 0)                AS out_zar_closed
+    COUNT(*)::int                                         AS out_referrers_paid,
+    COALESCE(SUM((c.total_zar * 100)::bigint), 0)::bigint AS out_points_awarded,
+    COALESCE(SUM(c.total_zar), 0)::numeric                AS out_zar_closed
   FROM closing_summary c;
 END;
 $$;
